Core and Module Security Updates DB UP reqd
Database update required

+++ web/modules/contrib/entity_share/src/EntityShareUtility.php
@@ -128,7 +128,7 @@ class EntityShareUtility {
       $channel_maxsize = $channel_infos[$channel_id]['channel_maxsize'];
     }
 
-    return min($import_maxsize, $channel_maxsize);
+    return (int) min($import_maxsize, $channel_maxsize);
   }
 
 }
